+++ src/safdom.js
@@ -4,7 +4,7 @@
 /**
 * PRIMARY TARGET IS LATEST CHROMIUM (Chrome 45+), shorthands and calls optimized for use with it
 * all other browsers compatibility notes are JUST FOR REFERENCE, and not for all features,
-* although latest versions of FF, Opera, Safari and even IE (Edge) can handle everything fine.
+* although latest versions of FF, Opera, Safari and even IE (Edge) can handle most things fine.
 * Feature       Chrome min version  Notes
 * $f            1+
 * $1            1+
@@ -19,7 +19,7 @@
 * $class*       8+                  Dependent on Element.classList
 */
 
-// Object.assign is not supported by IE w/o polyfill (but Edge has it), others: [Chrome 45+, FF 34+, Opera 32+, Safari 9+]
+// `Object.assign` is not supported by IE w/o polyfill (but Edge has it), others: [Chrome 45+, FF 34+, Opera 32+, Safari 9+]
 // however, the following IS NOT A POLYFILL, just a shorthand
 function _extendX(o) {
     for (var k, ai = 1; a = arguments[ai]; ++ai) for (k in a) if (a.hasOwnProperty(k)) o[k] = a[k];
@@ -27,6 +27,68 @@ function _extendX(o) {
 };
 function _extend(ctx) { return ctx.Object.assign || _extendX; };
 
+// internal utility methods:
+function _util_compatMode(ctx) {
+    return (ctx.document.compatMode || '') === 'CSS1Compat';
+};
+// scrollElement shim via https://github.com/mathiasbynens/document.scrollingElement, partially rewritten
+// see also: http://www.quirksmode.org/mobile/tableViewport.html
+function _util_mkScrollingElementGetter(ctx) {
+    var document = ctx.document;
+    if ('scrollingElement' in document) return function() { return document.scrollingElement; }; // woohoo, simple case
+    // Well, we need a runtime detector for that. So frustrating.
+    var isBodyElement = ctx.HTMLBodyElement
+        ? function(elt) { return elt instanceof HTMLBodyElement; }
+        : function(elt) { return elt.tagName.toLocaleLowerCase() == 'body'; }; // `tagName` is uppercase in HTML, but may be any-case in XML
+    var locateBodyElement = function(frameset) {
+        // We use this function to be correct per spec in case `document.body` is
+        // a `frameset` but there exists a later `body`. Since `document.body` is
+        // a `frameset`, we know the root is an `html`, and there was no `body`
+        // before the `frameset`, so we just need to look at siblings after the
+        // `frameset`.
+        var current = frameset;
+        do {
+            if (current.nodeType == 1 && isBodyElement(current)) return current;
+        } while (current = current.nextSibling);
+        return null; // nothing found
+    };
+    var isScrollingElementCompliant = function() {
+        // In quirks mode, the result is equivalent to the non-compliant standards mode behavior.
+        if (!_util_compatMode(ctx)) return false;
+        var iframe = document.createElement('iframe');
+        iframe.style.height = '1px';
+        (document.body || document.documentElement || document).appendChild(iframe);
+        var iframeDoc = iframe.contentWindow.document;
+        iframeDoc.write('<!DOCTYPE html><div style="height:9999em">x</div>');
+        iframeDoc.close();
+        var result = iframeDoc.documentElement.scrollHeight > iframeDoc.body.scrollHeight;
+        iframe.parentNode.removeChild(iframe);
+        isScrollingElementCompliant = function() { return _util_compatMode(ctx) && result; }; // further calls faster
+        return result;
+    };
+    var isScrollable = function(body) {
+        // a `body` element is scrollable if `body` and `html` both have non-`visible` overflow and are both being rendered
+        // about check that element is rendered:
+        // I've cut the visibility == 'collapse' test for table-*, it can be things-breaking for some setups
+        // or maybe do I check offsetHeight and offsetParent values instead messing with styles?
+        //TODO investigate that
+        var bodyStyle = elt.$cs(),
+            htmlStyle = document.documentElement.$cs();
+        return bodyStyle.overflow != 'visible' && bodyStyle.display != 'none' && bodyStyle.visibility == 'visible'
+            && htmlStyle.overflow != 'visible' && htmlStyle.display != 'none' && htmlStyle.visibility == 'visible';
+    };
+    return function() {
+        if (isScrollingElementCompliant()) return document.documentElement; // better than it can be
+        // Need duck typing body element and check other heavy shit. Every fucking time. R.I.P. performance.
+        // you should note that document.body could be not only a HTMLBodyElement,
+        // but at least HTMLFrameSetElement or null (e.g. if not yet initialized)
+        var body = document.body && locateBodyElement(document.body);
+        // If `body` is itself scrollable, it is not the `scrollingElement`.
+        return body && isScrollable(body) ? null : body;
+    };
+};
+
+// prototype extensions:
 function _protoNode(ctx) {
     var Node = ctx.Node,
         DPP = Node.DOCUMENT_POSITION_PRECEDING,
@@ -143,17 +205,61 @@ function _protoElement(ctx) {
     //TODO rename to $psa/$nsa or completely drop (too complicated)
     $pss: function() { return this.parentElement.querySelectorAll(':scope > :not(:nth-child(' + this.$idx() + ') ~ *)'); }, // all siblings from parentElement.firstElementChild to this.previousElementSibling => NONlive HTMLCollection
     $nss: function() { return this.querySelectorAll(':scope ~ *'); }, // all siblings from this.nextElementSibling, to parentElement.lastElementChild => NONlive HTMLCollection
+
+    // utility
+    // alias for getComputedStyle with the minimal "shim" SUDDENLY for IE < 9
+    $cs: ctx.getComputedStyle
+        ? function(pseudoElt) { return ctx.getComputedStyle(this, pseudoElt || null); }
+        : function(pseudoElt) { return this.currentStyle; },
+
+    // aliases for getBoundingClientRect/getClientRects
+    // take a look at https://msdn.microsoft.com/en-us/library/hh781509.aspx
+    $rect: protoElement.getBoundingClientRect,
+    $rects: protoElement.getClientRects,
 }; };
 
 function _protoEventTarget(ctx) {
     return {
-    // EventTarget shorthands
+    // those are shorthands for objects that implement EventTarget interface
+    // however, EventTarget.prototype itself may not be available, see $SAF()
     //TODO
     $on: function(event, handler) { return this.addEventListener(event, handler, false), this; },
     $onc: function(event, handler) { return this.addEventListener(event, handler, true), this; },
     $off: function(event, handler) { return this.removeEventListener(event, handler, false), this; },
     $offc: function(event, handler) { return this.removeEventListener(event, handler, true), this; },
 }; };
+
+// browser/window/context/viewport/document utils/stats
+function _meta(ctx) {
+    var extend = _extend(ctx),
+        getSE = _util_mkScrollingElementGetter(ctx);
+        meta = {
+            se: getSE,
+            slG: function() { return getSE().scrollLeft; }, // scrollLeft
+            slS: function(scrollLeft) { return (getSE().scrollLeft = scrollLeft), meta; },
+            stG: function() { return getSE().scrollTop; }, // scrollTop
+            stS: function(scrollTop) { return (getSE().scrollTop = scrollTop), meta; },
+            vw: function() { return ctx.innerWidth; }, // viewport width (disregarding vertical scrollbar)
+            vh: function() { return ctx.innerHeight; }, // viewport height (disregarding horisontal scrollbar)
+            cw: function() { return getSE().clientWidth; }, // client width (vw - vertical scrollbar width, if present)
+            ch: function() { return getSE().clientHeight; }, // client height (vh - horisontal scrollbar width, if present)
+            sw: function() { return getSE().scrollWidth; }, // scroll width
+            sh: function() { return getSE().scrollHeight; }, // scroll height
+        };
+
+    // pageXYoffset to use with $rect(s)
+    // see https://developer.mozilla.org/en-US/docs/Web/API/Window/scrollY
+    if ('pageXOffset' in ctx) extend(meta, {
+        px: function() { return ctx.pageXOffset; },
+        py: function() { return ctx.pageYOffset; },
+    });
+    else extend(meta, {
+        px: function() { return getSE().scrollLeft; },
+        py: function() { return getSE().scrollTop; },
+    });
+
+    return meta;
+};
 
 function $SAF(ctx) { // extend ctx with SAFDOM methods
     var extend = _extend(ctx),
@@ -164,7 +270,19 @@ function $SAF(ctx) { // extend ctx with SAFDOM methods
     extend(ctx.NodeList.prototype, protoArr);
     extend(ctx.HTMLCollection.prototype, protoArr);
     extend(ctx.Element.prototype, _protoElement(ctx));
-    extend(ctx.EventTarget.prototype, _protoEventTarget(ctx));
+
+    // wtf, ie? https://developer.mozilla.org/en-US/docs/Web/API/Event/target
+    // so, EventTarget.prototype itself may not be available, therefore we need to extend individual things
+    var protoET = _protoEventTarget(ctx);
+    if (ctx.EventTarget) {
+        extend(ctx.EventTarget.prototype, protoET);
+    }
+    else {
+        // touch main things, leave others like XMLHttpRequest unmodified
+        extend(ctx.self, protoET);
+        extend(ctx.document, protoET);
+        extend(ctx.Node.prototype, protoET);
+    }
 
     return extend(ctx, {
     // Globals
@@ -177,6 +295,7 @@ function $SAF(ctx) { // extend ctx with SAFDOM methods
     $mkT: document.createTextNode.bind(document),
     $mkF: document.createDocumentFragment.bind(document),
     //TODO $mk*, $mkEvent
+    $meta: _meta(ctx),
     });
 };